+++ packages/core/src/metas/initialize.test.ts
@@ -2,6 +2,14 @@ import { initializeFaro } from '../initialize';
 import { mockConfig } from '../testUtils';
 
 describe('metas', () => {
+  beforeAll(() => {
+    delete (global as any).FARO_BUILD_ID_TEST;
+  });
+
+  afterAll(() => {
+    delete (global as any).FARO_BUILD_ID_TEST;
+  });
+
   it('can set listeners and they will be notified on meta changes', () => {
     const { metas } = initializeFaro(mockConfig());
 
@@ -18,5 +26,12 @@ describe('metas', () => {
 
     metas.add({ session: { id: '2' } });
     expect(listener).toHaveBeenCalledTimes(2);
+  });
+
+  it('can get the build ID from the global object', () => {
+    (global as any).FARO_BUILD_ID_TEST = 'fizzbuzz';
+    const { metas: { value: { app } } } = initializeFaro(mockConfig());
+
+    expect(app?.buildId).toEqual('fizzbuzz');
   });
 });