+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Next
 
+- Deps (`@grafana/faro-web-tracing`, `@grafana/faro-core`): Update OpenTelemetry dependencies (#475).
+
 ## 1.3.7
 
 - Enhancement (`@grafana/faro-web-sdk`): add response time to performance timings (#465).